requirements.txt
langchain
langchain-openai
langchain-community
langchain-core
openai
chromadb
pandas
geopandas
matplotlib
folium
shapely
numpy
python-dotenv
